tests/wpt/web-platform-tests/cookies/domain/domain-attribute-matches-host.sub.https.html

<!DOCTYPE html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/cookies/resources/cookie-helper.sub.js"></script>
</head>
<body>
  <script>
    //
    //  Set-Cookie: domain-attribute-matches-host=b; Path=/; Domain={{host}}
    //
    const cookieName = "domain-attribute-matches-host";
    test(t => {
      assert_dom_cookie(cookieName, "b", true);
    }, "Domain={{host}} => available via `document.cookie`");

    async_test(t => {
      fetch("/cookies/resources/list.py", { credentials: "include" })
        .then(t.step_func(r => r.json()))
        .then(t.step_func_done(r => {
          assert_equals(r[cookieName], "b");
        }))
        .catch(_ => assert_unreached);
    }, "Domain={{host}} => sent with same-origin requests.");

    async_test(t => {
      fetch(`${SECURE_SUBDOMAIN_ORIGIN}/cookies/resources/list.py`, { credentials: "include" })
        .then(t.step_func(r => r.json()))
        .then(t.step_func_done(r => {
          assert_equals(r[cookieName], "b");
        }))
        .catch(_ => assert_unreached);
    }, "Domain={{host}} => sent with subdomain requests.");
  </script>
</body>
