LayoutTests/imported/w3c/web-platform-tests/cookies/domain/domain-attribute-matches-host.sub.https-expected.txt

Blocked access to external URL https://www1.localhost:9443/cookies/resources/list.py
CONSOLE MESSAGE: Fetch API cannot load https://www1.localhost:9443/cookies/resources/list.py due to access control checks.

Harness Error (TIMEOUT), message = null

PASS Domain=localhost => available via `document.cookie`
PASS Domain=localhost => sent with same-origin requests.
TIMEOUT Domain=localhost => sent with subdomain requests. Test timed out
